+++ Image FeedTests/ImageViewTests.swift
@@ -34,7 +34,6 @@ final class ImagesListViewControllerTests: XCTestCase {
     // MARK: - ImagesListViewController Tests
     
     func testUpdateTableViewAnimated_NoChangeInPhotoCount() {
-        // Arrange
         let photo = Photo(
             id: "1",
             size: CGSize(width: 100, height: 100),
@@ -48,17 +47,15 @@ final class ImagesListViewControllerTests: XCTestCase {
         service.photos = [photo]
         viewController.updateTableViewAnimated()
         
-        // Act
         tableView.reset()
         service.photos = [photo]
         viewController.updateTableViewAnimated()
         
-        // Assert
         XCTAssertFalse(tableView.insertRowsCalled, "Expected no rows to be inserted")
     }
     
     func testUpdateTableViewAnimated_NewPhotosAdded() {
-        // Arrange
+
         let initialPhoto = Photo(
             id: "1",
             size: CGSize(width: 100, height: 100),
@@ -82,12 +79,10 @@ final class ImagesListViewControllerTests: XCTestCase {
         service.photos = [initialPhoto]
         viewController.updateTableViewAnimated()
         
-        // Act
         tableView.reset()
         service.photos.append(newPhoto)
         viewController.updateTableViewAnimated()
         
-        // Assert
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
             XCTAssertTrue(self.tableView.insertRowsCalled, "Expected rows to be inserted")
             XCTAssertEqual(self.tableView.insertRowsIndexPaths.count, 1, "Expected 1 row to be inserted")
@@ -98,15 +93,12 @@ final class ImagesListViewControllerTests: XCTestCase {
     }
     
     func testConfigureCell() {
-        // given
         let cell = TestImagesListCell()
         let indexPath = IndexPath(row: 0, section: 0)
         service.photos = [Photo(id: "1", size: CGSize(width: 100, height: 100), createdAt: nil, welcomeDescription: nil, thumbImageURL: "", largeImageURL: "", isLiked: false)]
         
-        // when
         presenter.configureCell(cell, at: indexPath)
         
-        // then
         XCTAssertNotNil(cell.customImageView.image)
     }
     
@@ -126,15 +118,12 @@ final class ImagesListViewControllerTests: XCTestCase {
     }
     
     func testTableViewHeightForRowAt() {
-        // given
         let indexPath = IndexPath(row: 0, section: 0)
         let photo = Photo(id: "1", size: CGSize(width: 100, height: 200), createdAt: nil, welcomeDescription: nil, thumbImageURL: "", largeImageURL: "", isLiked: false)
         viewController.photos = [photo]
         
-        // when
         let height = viewController.tableView(viewController.tableView, heightForRowAt: indexPath)
         
-        // then
         let expectedHeight = viewController.tableView.frame.width * photo.size.height / photo.size.width
         XCTAssertEqual(height, expectedHeight, accuracy: 0.01)
     }
@@ -142,43 +131,32 @@ final class ImagesListViewControllerTests: XCTestCase {
     // MARK: - ImagesListPresenter Tests
     
     func testNumberOfRows() {
-        // given
         service.photos = [Photo(id: "1", size: CGSize(width: 100, height: 100), createdAt: nil, welcomeDescription: nil, thumbImageURL: "", largeImageURL: "", isLiked: false)]
         
-        // when
         let rows = presenter.numberOfRows()
         
-        // then
         XCTAssertEqual(rows, 1)
     }
     
     func testDidSelectRow() {
-        // given
         let indexPath = IndexPath(row: 0, section: 0)
         
-        // when
         presenter.didSelectRow(at: indexPath)
         
-        // then
         XCTAssertTrue(presenter.performSegueCalled)
     }
     
     func testFetchNextPage() {
-        // when
         presenter.fetchNextPage()
         
-        // then
         XCTAssertTrue(service.fetchNextPageCalled)
     }
     
     func testHandlePhotosUpdated() {
-        // given
         service.photos = [Photo(id: "1", size: CGSize(width: 100, height: 100), createdAt: nil, welcomeDescription: nil, thumbImageURL: "", largeImageURL: "", isLiked: false)]
         
-        // when
         presenter.handlePhotosUpdated()
         
-        // then
         XCTAssertTrue(presenter.updateTableViewAnimatedCalled)
     }
 }
